requirements.txt
appdirs==1.4.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
docker==6.1.3
docker-pycreds==0.4.0
filelock==3.12.2
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.74
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
wandb==0.15.8
websocket-client==1.6.1
Werkzeug==2.3.6
